+++ NEWS.md
@@ -1,6 +1,8 @@
 # resourcecode (development version)
 
-- Adds`{resourcecodedata}` package as dependence, to ease the use of the package 
+- New function `metconv2zmcomp()` to compute the u/v components
+  from meteorological wind speed and direction.
+- Adds`{resourcecodedata}` package as dependence, to ease the use of the package .
 - `closest_point_field()` and `closest_point_field()` now accept a list of locations
   instead of a single one (#3).